jQuery Scala Stack - Renaming a lot! - "Catch all filter" excerpting "/api" calls - More packages
